+++ packages/plugins/vite-plugin-generate-shopify-locales/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @solidifront/vite-plugin-generate-shopify-locales
 
+## 1.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @solidifront/storefront-client@0.2.2
+
+## 1.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @solidifront/storefront-client@0.2.1
+
 ## 1.1.3
 
 ### Patch Changes